+++ doc/documents/changelog.md
@@ -1,3 +1,8 @@
 # Change Log
+**May 11th, 2020**
+- Added Error Message to Login callback
+- profile page
+- API for email update is made private `/api/v1/user/email`
+
 **May 10th, 2020**
 - Added API call to grab user information `/api/v1/user`. Learn more in Developer Documentation.